requirements.txt
Django==3.0.3
whitenoise==5.0.1
gunicorn
psycopg2==2.8.4
psycopg2-binary
djangorestframework
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
django-model-utils
BeautifulSoup4
requests
python-social-auth
social-auth-app-django
social-auth-core